+++ src/matterbridgeDeviceTypes.ts
@@ -183,12 +183,76 @@ export const DeviceTypeDefinition = ({
   optionalClientClusters,
 });
 
+/**
+ *  MatterbridgeEndpointOptions interface is used to define the options for a Matterbridge endpoint.
+ *
+ *  @remarks
+ *  - tagList?: Semtag[]. It is used to disambiguate the sibling child endpoints (9.2.3. Disambiguation rule).
+ *    - mfgCode: VendorId | null,
+ *    - namespaceId: number,
+ *    - tag: number,
+ *    - label: string | undefined | null
+ *  - mode?: 'server' | 'matter'. It is used to activate a special mode for the endpoint.
+ *  - id?: string. It is the unique storage key for the endpoint.
+ *  - number?: EndpointNumber. It is the endpoint number for the endpoint.
+ */
 export interface MatterbridgeEndpointOptions extends EndpointOptions {
+  /**
+   *  The semantic tags array for the endpoint.
+   *  The tagList is used to disambiguate the sibling child endpoints (9.2.3. Disambiguation rule).
+   *  The tagList is used to identify the endpoint and to provide additional information about the endpoint.
+   *
+   *  @remarks
+   *    - mfgCode: VendorId | null,
+   *    - namespaceId: number,
+   *    - tag: number,
+   *    - label: string | undefined | null
+   */
   tagList?: Semtag[];
+  /**
+   * Activates a special mode for this endpoint.
+   * - 'server': it creates the device server node and add the device as Matter device that needs to be paired individually.
+   *   In this case the bridge mode is not relevant. The device is autonomous.
+   *
+   * - 'matter': it adds the device directly to the bridge server node as Matter device. In this case the implementation must respect
+   *   the 9.2.3. Disambiguation rule (i.e. use taglist if needed cause the device doesn't have nodeLabel).
+   *   Furthermore the device will be a part of the bridge (i.e. will have the same name and will be in the same room).
+   *   See 9.12.2.2. Native Matter functionality in Bridge.
+   *
+   * @remarks
+   * Always use createDefaultBasicInformationClusterServer() to create the BasicInformation cluster server.
+   */
+  mode?: 'server' | 'matter';
+  /**
+   * The unique storage key for the endpoint.
+   * If not provided, a default key will be used.
+   */
+  id?: string;
+  /**
+   * The endpoint number for the endpoint.
+   * If not provided, the endpoint will be created with the next available endpoint number.
+   * If provided, the endpoint will be created with the specified endpoint number.
+   */
+  number?: EndpointNumber;
 }
 
+/**
+ *  EndpointOptions interface is used to define the options for an endpoint.
+ *
+ * @deprecated Use `MatterbridgeEndpointOptions` instead.
+ */
 export interface EndpointOptions {
+  /**
+   * Old API compatibility replaced by number
+   *
+   * @deprecated Use `number` instead.
+   */
   endpointId?: EndpointNumber;
+  /**
+   * Old API compatibility replaced by id
+   *
+   * @deprecated Use `id` instead.
+   */
   uniqueStorageKey?: string;
 }